+++ src/main/java/zdoctor/morecrops/zcore/common/EasyBlock.java
@@ -1,5 +1,7 @@
-package zcore.common;
+package zdoctor.morecrops.zcore.common;
 
+import net.minecraft.block.Block;
+import net.minecraft.block.material.Material;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.resources.model.ModelResourceLocation;
 import net.minecraft.creativetab.CreativeTabs;
@@ -10,51 +12,56 @@ import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.registry.GameRegistry;
 import net.minecraftforge.fml.relauncher.Side;
-import zcore.proxy.CommonProxy;
+import zdoctor.morecrops.zcore.proxy.CommonProxy;
+
 /**
  * A class to automate the process of putting it into the game and getting its texture
  * @author Z_Doctor
  */
-public class EasyItem extends Item implements ISubEvent {
+public class EasyBlock extends Block implements ISubEvent {
 	// Basic info
-	protected final String itemModel;
+	protected final String blockModel;
 	protected final String modID;
 	// Advance
 	protected Object[] recipe;
 	protected boolean isShapeless;
 	// Constructors
-	public EasyItem(String model, String mod) {
-		this(model, mod, CreativeTabs.tabMisc);
+	public EasyBlock(String model, String mod) {
+		this(model, mod, CreativeTabs.tabDecorations);
 	}
-	public EasyItem(String model, String mod, CreativeTabs tab) {
-		this.itemModel = model;
+	public EasyBlock(String model, String mod, CreativeTabs tab) {
+		super(Material.rock);
+		this.blockModel = model;
 		this.modID = mod;
 		
 		this.setCreativeTab(tab);
 		this.setUnlocalizedName(this.getModelPath());
+		this.setHardness(1.5F);
+		this.setResistance(10.0F);
+		this.setStepSound(soundTypePiston);
 		
 		CommonProxy.subEvent(this);	
 	}
 	
 	/** Override to set a different path */
 	public String getModelPath() {
-		return this.itemModel;
+		return this.blockModel;
 	}
 	/** Override to change meta */
-	public int getItemMeta() {
+	public int getBlockMeta() {
 		return 0;
 	}
 	
 	/** Shaped by default 
 	 * @param recipe - The food's recipe object */
-	public EasyItem setRecipe(Object[] recipe) {
+	public EasyBlock setRecipe(Object[] recipe) {
 		return this.setRecipe(recipe, false);
 	}
 	/**
 	 * @param recipe - The item's recipe
 	 * @param isShapeless - Whether or not the food's recipe is shapeless
 	 */
-	public EasyItem setRecipe(Object[] recipe, boolean isShapeless) {
+	public EasyBlock setRecipe(Object[] recipe, boolean isShapeless) {
 		if(this.recipe == null) {
 			this.recipe = recipe;
 			this.isShapeless = isShapeless;
@@ -62,33 +69,30 @@ public class EasyItem extends Item implements ISubEvent {
 		return this;
 	}
 	
-	public EasyItem setNoTab() {
+	public Item asItem() {
+		return Item.getItemFromBlock(this);
+	}
+	public EasyBlock setNoTab() {
 		this.setCreativeTab((CreativeTabs)null);
 		return this;
 	}
 	
 	protected void registerRender() {
-		System.out.println(this.modID + ":" + this.getModelPath());
-		Minecraft.getMinecraft().getRenderItem().getItemModelMesher().register(this, this.getItemMeta(), 
+		Minecraft.getMinecraft().getRenderItem().getItemModelMesher().register(this.asItem(), this.getBlockMeta(), 
 				new ModelResourceLocation(this.modID + ":" + this.getModelPath(), "inventory"));
-	}
-	
-	public boolean registerRecipe() {
-		return true;
 	}
 	
 	// Overrides
 	@Override
 	public void fire(FMLPreInitializationEvent e) {
-		System.out.println("Registering: " + this.modID + ":" + this.getModelPath());
-		GameRegistry.registerItem(this, this.getModelPath(), this.modID);
-		
+		System.out.println(this.modID + ":" + this.getModelPath());
+		GameRegistry.registerBlock (this, this.getModelPath());
 	}
 	@Override
 	public void fire(FMLInitializationEvent e) {
 		if(e.getSide() == Side.CLIENT)
-			registerRender();
-		if(this.recipe != null && this.registerRecipe()) {
+			this.registerRender();
+		if(this.recipe != null) {
 			if(this.isShapeless)
 				GameRegistry.addShapelessRecipe(new ItemStack(this), this.recipe);
 			else
@@ -97,4 +101,9 @@ public class EasyItem extends Item implements ISubEvent {
 	}
 	@Override
 	public void fire(FMLPostInitializationEvent e) {}
+	/** Override to change material */
+	@Override
+	public Material getMaterial() {
+		return super.getMaterial();
+	}
 }