+++ src/main/java/zdoctor/morecrops/zcore/common/EasyStuff.java
@@ -1,10 +1,17 @@
-package zcore.common;
+package zdoctor.morecrops.zcore.common;
 
+import java.util.Iterator;
+
+import net.minecraft.init.Items;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.crafting.CraftingManager;
+import net.minecraft.item.crafting.IRecipe;
 import net.minecraft.util.WeightedRandomChestContent;
 import net.minecraftforge.common.ChestGenHooks;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.FMLCommonHandler;
+import net.minecraftforge.fml.common.registry.GameRegistry;
 
 public class EasyStuff {
 	public static class EventRegistry {
@@ -72,4 +79,52 @@ public class EasyStuff {
 			ChestGenHooks.addItem(category, new WeightedRandomChestContent(iStack, min, max, rariry));
 		}
 	}
+	/**
+	 * A collection of functions for quick game development
+	 * @author Z_Doctor
+	 */
+	public static class GameModifications {
+		/**
+		 * Removes given item(s) from CraftingManager's Recipe List
+		 * @param itemList - A single or a collection of items to be removed (item1, item2, item3, etc)
+		 */
+		public static void removeItemRecipe(Item...itemList) {
+			for(Item item : itemList) {
+				Iterator<IRecipe> RecipeRegistry = CraftingManager.getInstance().getRecipeList().iterator();
+				while (RecipeRegistry.hasNext()){
+					IRecipe recipe = RecipeRegistry.next();
+					if(recipe == null)
+						continue;
+					ItemStack output = recipe.getRecipeOutput();
+					if(output != null && item == output.getItem())
+						RecipeRegistry.remove();
+				}
+			}
+		}
+	
+		public static void registerItemRecipe(Item item, Object[] recipe, boolean shapeless) {
+			if(item != null && recipe != null) {
+				if(shapeless)
+					GameRegistry.addShapelessRecipe(new ItemStack(item), recipe);
+				else
+					GameRegistry.addRecipe(new ItemStack(item), recipe);
+			}
+		}
+	}
+
+	public static class RegistryChecker {
+		public static boolean checkForRecipe(Item item) {
+			Iterator<IRecipe> RecipeRegistry = CraftingManager.getInstance().getRecipeList().iterator();
+			while (RecipeRegistry.hasNext()){
+				IRecipe recipe = RecipeRegistry.next();
+				if(recipe == null)
+					continue;
+				ItemStack output = recipe.getRecipeOutput();
+				if(output != null && item == output.getItem())
+					return true;
+			}
+			return false;
+		}
+	}
+
 }