+++ ChangeLog.md
@@ -1,3 +1,16 @@
+1.4
+    - Added Achievements
+    - Plants now only drop one seed with a chance of dropping another seed
+    - Changed Configuration (again)
+        - Can now disable each seed recipe individually (no longer requires restart to change)
+        - Added some Config Welcome Messages
+        - Can now control the drop rate of each seed
+    - Code a bit more organized
+    - Added Glowstone and Nether Quartz seeds (only growable in nether)
+    - Seeds (except Diamond Seeds) now only grow in nether (can change in config).
+    - Seeds do not grow in peacful mode (config item)
+    - No more food items will be added to this mod, because I will make that a seperate mod.
+
 1.3
     - Added Strawberry Crop, Seed and Food Item (More on the way)
         - Has textures (Beautiful)